requirements.txt
streamlit
pandas
openpyxl
groq
python-dotenv
scikit-learn